Fix: FileNotFoundError
By default, the Zig compiler outputs the binary executable and the
shared object file created with `zig build-exe FILE` to the current
working directory, and not where the source file is run.

Changed `zig_build_exe()` so that the current working directory is
changed to the project's root directory before running `zig build-exe`.
This, also, simplifies things when calling `Path.replace()` to move the
outputted files into their respective directories.

The current working directory (from where the script was run) is saved
and changed back to after the command is run.

As I'm typing this commit message, I've just realized I don't need to do
`os.chdir(cwd)` for each `try-except-else-finally` block. It just needs
to be in the `finally` block.